Generate sources at configure time.
The example illustrate how to generate a simple source file that defines
a function. The example is from "The CMake Cookbook", recipe 01 in
chapter 6Generate sources at configure time.

The example illustrate how to generate a simple source file that defines
a function. The example is from "The CMake Cookbook", recipe 01 in
chapter 6.